WIP move items to be functions
